requirements.txt
apache-airflow==2.9.2
pandas==2.0.0
numpy==1.26.4
configparser
boto3
kagglehub
glob2